Extract PropertyHookThrowPointsResolver from NodeScopeResolver
